refactor: extract conversation custom attributes actions into concern
Removes the Metrics/ClassLength rubocop disable by moving the
custom_attributes and destroy_custom_attributes actions into a
dedicated concern, following the existing LabelConcern pattern.
