+++ app/controllers/api/v1/accounts/conversations_controller.rb
@@ -1,7 +1,8 @@
-class Api::V1::Accounts::ConversationsController < Api::V1::Accounts::BaseController # rubocop:disable Metrics/ClassLength
+class Api::V1::Accounts::ConversationsController < Api::V1::Accounts::BaseController
   include Events::Types
   include DateRangeHelper
   include HmacConcern
+  include ConversationCustomAttributesConcern
 
   before_action :conversation, except: [:index, :meta, :search, :create, :filter]
   before_action :inbox, :contact, :contact_inbox, only: [:create]
@@ -127,20 +128,6 @@ class Api::V1::Accounts::ConversationsController < Api::V1::Accounts::BaseContro
     last_incoming_message = @conversation.messages.incoming.last
     last_seen_at = last_incoming_message.created_at - 1.second if last_incoming_message.present?
     update_last_seen_on_conversation(last_seen_at, true)
-  end
-
-  def custom_attributes
-    attributes = params.permit(custom_attributes: {})[:custom_attributes]
-    # When `merge` is truthy, only the keys sent are updated and the rest are kept, matching the contacts endpoint.
-    # Replace stays the default so existing integrations are unaffected.
-    attributes = @conversation.custom_attributes.merge(attributes || {}) if ActiveModel::Type::Boolean.new.cast(params[:merge])
-    @conversation.custom_attributes = attributes
-    @conversation.save!
-  end
-
-  def destroy_custom_attributes
-    @conversation.custom_attributes = @conversation.custom_attributes.excluding(params[:custom_attributes])
-    @conversation.save!
   end
 
   def destroy
